refactor(search): replace local Nostr schemas with zod-nostr
