+++ package.json
@@ -51,6 +51,7 @@
     "nostr-tools": "^2.23.9",
     "rx-nostr": "^3.7.5",
     "rxjs": "^7.8.2",
-    "zod": "^4.4.3"
+    "zod": "^4.4.3",
+    "zod-nostr": "^0.1.1"
   }
 }
