+++ src/lib/search/result.ts
@@ -1,6 +1,19 @@
 import type * as Nostr from 'nostr-typedef';
 import { z } from 'zod';
-import { NostrEventSchema } from './nostr';
+import { zostr } from 'zod-nostr';
+
+export const VerifiedNostrEventSchema = zostr
+  .event()
+  .check(zostr.signatureCheck())
+  .transform(({ id, pubkey, created_at, kind, tags, content, sig }) => ({
+    id,
+    pubkey,
+    created_at,
+    kind,
+    tags,
+    content,
+    sig,
+  }));
 
 export const SearchResultPaginationSchema = z.object({
   last_page: z.boolean(),
@@ -13,7 +26,7 @@ export const SearchResultPaginationSchema = z.object({
 
 export const SearchResultSchema = z.object({
   // Requests are capped at 100 results, so cap verification work to the same bound.
-  data: z.array(NostrEventSchema).max(100),
+  data: z.array(VerifiedNostrEventSchema).max(100),
   pagination: SearchResultPaginationSchema,
 });
